Implement stack backtrace Implement Lagrange interpolation

+++ tests/cpp/test_copy_from_segmented_array.cpp
@@ -15,13 +15,13 @@ int main(void) {
     // [1.0, 5.0, 9.0 ]
     // [2.0, 6.0, 10.0]
 
-     merlin::array::Array X({1024,256,8});
+    merlin::array::Array X({1024,256,8});
 
     MESSAGE("Initialize Array A.\n");
-    float A[10] = {1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0};
+    double A[10] = {1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0};
     std::uint64_t ndim = 2;
     std::uint64_t dims[2] = {3, 2};
-    std::uint64_t strides[2] = {2*(dims[1] * sizeof(float)), sizeof(float)};
+    std::uint64_t strides[2] = {2*(dims[1] * sizeof(double)), sizeof(double)};
 
     // copy array
     MESSAGE("Copy Array A to Ar_copy.\n");
